% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_values.R
\name{calculate_values}
\alias{calculate_values}
\title{Sum values}
\usage{
calculate_values(num)
}
\arguments{
\item{num}{a number}
}
\value{
2 times num plus 4
}
\description{
Sum values
}
\examples{
calculate_values(3)
}
