% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nothing.R
\name{do_nothing}
\alias{do_nothing}
\title{Do nothing function}
\usage{
do_nothing(a)
}
\arguments{
\item{a}{can be anything. It does nothing with it.}
}
\value{
nothing
}
\description{
It does nothing.
}
\examples{
do_nothing()
}
