% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_something.R
\name{do_something}
\alias{do_something}
\title{Do something for Jasper's package}
\usage{
do_something(name)
}
\arguments{
\item{name}{What is your name?}
}
\value{
A greeting message
}
\description{
Do something for Jasper's package
}
\examples{
do_something("Jasper")
}
