% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_something.R
\name{do_something}
\alias{do_something}
\title{Do something}
\usage{
do_something(n = 5)
}
\arguments{
\item{n}{an integer, how much do you want?}
}
\value{
your something (a vector of standard normal numbers)
}
\description{
Do something
}
\examples{
do_something(42)
}
